Move _search_for_svr_hyperparams() into tuning.py
